requirements.txt
pytest==7.4.4
python-telegram-bot==22.0